backend/requirements.txt
# backend/requirements.txt
flask
python-dotenv
google-generativeai
requests 